[DEITS] Rollback import after failure

+++ packages/core/data-transfer/src/strapi/providers/local-destination/index.ts
@@ -1,7 +1,13 @@
 import { Writable } from 'stream';
 import path from 'path';
 import * as fse from 'fs-extra';
-import type { IAsset, IDestinationProvider, IMetadata, ProviderType } from '../../../../types';
+import type {
+  IAsset,
+  IDestinationProvider,
+  IMetadata,
+  ProviderType,
+  Transaction,
+} from '../../../../types';
 
 import { restore } from './strategies';
 import * as utils from '../../../utils';
@@ -25,6 +31,8 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
 
   strapi?: Strapi.Strapi;
 
+  transaction?: Transaction;
+
   /**
    * The entities mapper is used to map old entities to their new IDs
    */
@@ -38,10 +46,13 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
   async bootstrap(): Promise<void> {
     this.#validateOptions();
     this.strapi = await this.options.getStrapi();
+
+    this.transaction = utils.transaction.createTransaction(this.strapi);
   }
 
   async close(): Promise<void> {
     const { autoDestroy } = this.options;
+    this.transaction?.end();
 
     // Basically `!== false` but more deterministic
     if (autoDestroy === undefined || autoDestroy === true) {
@@ -63,10 +74,24 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
     return restore.deleteRecords(this.strapi, this.options.restore);
   }
 
+  async rollback(e: Error): Promise<void> {
+    await this.transaction?.rollback();
+  }
+
   async beforeTransfer() {
-    if (this.options.strategy === 'restore') {
-      await this.#deleteAll();
+    if (!this.strapi) {
+      throw new Error('Strapi instance not found');
     }
+
+    await this.transaction?.attach(async () => {
+      try {
+        if (this.options.strategy === 'restore') {
+          await this.#deleteAll();
+        }
+      } catch (error) {
+        throw new Error(`restore failed ${error}`);
+      }
+    });
   }
 
   getMetadata(): IMetadata {
@@ -120,6 +145,7 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
       return restore.createEntitiesWriteStream({
         strapi: this.strapi,
         updateMappingTable,
+        transaction: this.transaction,
       });
     }
 
@@ -183,7 +209,7 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
     const { strategy } = this.options;
 
     if (strategy === 'restore') {
-      return restore.createConfigurationWriteStream(this.strapi);
+      return restore.createConfigurationWriteStream(this.strapi, this.transaction);
     }
 
     throw new Error(`Invalid strategy supplied: "${strategy}"`);
@@ -198,7 +224,7 @@ class LocalStrapiDestinationProvider implements IDestinationProvider {
     const mapID = (uid: string, id: number): number | undefined => this.#entitiesMapper[uid]?.[id];
 
     if (strategy === 'restore') {
-      return restore.createLinksWriteStream(mapID, this.strapi);
+      return restore.createLinksWriteStream(mapID, this.strapi, this.transaction);
     }
 
     throw new Error(`Invalid strategy supplied: "${strategy}"`);
